6 submission

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -3,7 +3,7 @@ package lang.print.gaps.finalModuleTask;
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
 
-        int Fahrenheit;
+        float Fahrenheit;
         Fahrenheit =((temperatureCelsius*9)/5)+32;
         System.out.println( Fahrenheit);
     }
